+++ src/EventListener/ReaderBeforeRenderEventListener.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2020 Heimrich & Hannot GmbH
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -26,6 +26,7 @@ class ReaderBeforeRenderEventListener
     public function addNewsPagination(ReaderBeforeRenderEvent $event)
     {
         $context = (object) $event->getTemplateData();
+
         $this->manager->addNewsPagination($context, $event->getItem()->getFormatted(), $event->getReaderConfig());
 
         $event->setTemplateData((array) $context);